+++ packages/devtools/cli/src/daemon/index.ts
@@ -2,6 +2,5 @@
 // Copyright 2023 DXOS.org
 //
 
-export * from './pm2';
 export * from './run';
 export * from './util';
